Set IP address as master hostname for Vagrant
A Vagrant deployment was setting the `application.baseurl` property as
the master canonical hostname. Unfurtunately, such URL is not accessible
from the host server due to `master.vagrant` domain is not found.

This change introduces a condition while generating the property mentioned
above. For Vagrant, the IP address of master node is used. Otherwise, the
concat of hostname and domain is used as before.

This fixes the issue #1118