Create a Python library from cosmos-cli
As reported in issue #983, the code of cosmos-cli should be refactored
in order to provide a Python library. This change extracts the compute
management features of the CLI into a set of classes that may be reused
in other python projects.

* `cosmos.compute.protocol`, which provides a `Protocol` object that
encapsulates the API operations.
* `cosmos.compute.connection` and `cosmos.compute.cluster`, which provides
a set of convenience classes that makes the API more easy to use.