Bump up timeout value to avoid issues in CI server
The test class `LibVirtInfrastructureProviderTest` is failing from
time to time in `cosmos10` due to future timeouts. Any attempt to
reproduce the error out of CI environment was unsuccessful.

We are suspecting the timeout of 1 second is fine for our laptops but
may be not enough for cosmos10 when there is a high CPU load. If we
consider that there might be several Jenkins jobs running in parallel,
and each test class executes in parallel as well — the possibility of
timeout after 1 second is not absurd.

This change bumps up the timeout from 1 to 5 seconds. If there is
actually a deadlock, this won't mitigate the failures at all. Otherwise
the hypothesis described above would be confirmed.