Assert kernel version instead of install
The current deployment scripts attempts to install a concrete version
of OpenVZ kernel. If there is an older version of the kernel already
installed, the script fails, as described in issue #1175.

With this change, the Puppet script just executes a command to check
whether the kernel release is the one expected in the deployment.
If not, the deployment fails.