Implement CLI commands to add/rm cluster users
This change implements a portion of feature #1317 by adding the
necessary commands to Cosmos CLI in order to add/remove users to/from
a cluster. It stands on the following REST API operations still
unimplemented:

* `POST /api/cluster/<cluster_id>/add_user`, which adds a new user
to an existing cluster.
* `POST /api/cluster/<cluster_id>/remove_user`, which removes a user
from an existing cluster.

Both operations receive a JSON object comprised by a single field
`user` of string type, indicating the ID of the user to be added or
removed.