Remove some arguments from user registration
The user registration use case determines that properties as user group
and quota are not defined by the user but by the platform admin.

Therefore, there are some arguments in `CosmosProfileDao#registerUser()`
that should not come from the user registration process. This change
removes these arguments and replaces them by default properties declared
in the `CosmosProfileDao` and implemented in `DefaultUserProperties`.
Two additional method have been added to the DAO to get and set the
group for a given user.