Add user capabilities to DAO
As part of #1317, it is necessary to store in Cosmos API DAO whether
an user is capable of execute `sudo` command in his clusters. This is
done by considering a new property of `CosmosProfile` known as
`UserCapabilities`, which describes what special actions can be done
by the user on the platform resources.

This change performs the necessary modifications on `CosmosProfileDao`
in order to manage user capabilities.